pyproject.toml
[project]
name = "rock-club"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "blacksheep>=2.0.7",
    "mypy-baseline>=0.7.2",
    "mypy>=1.13.0",
    "piccolo[all]>=1.22.0",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
    "pyjwt>=2.10.0",
    "ruff>=0.8.0",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
]

[tool.mypy]
mypy_path = ["./src"]
